refactor: extract user menu
