+++ packages/frontend/editor-ui/src/components/MainSidebar.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { computed, onBeforeUnmount, onMounted, ref, nextTick, type Ref } from 'vue';
+import { computed, onBeforeUnmount, onMounted, ref, nextTick, type Ref, useTemplateRef } from 'vue';
 import { onClickOutside, type VueInstance } from '@vueuse/core';
 
 import { useI18n } from '@n8n/i18n';
@@ -13,7 +13,6 @@ import {
 	N8nLogo,
 	N8nPopoverReka,
 	N8nScrollArea,
-	N8nAvatar,
 	N8nText,
 	N8nIcon,
 	N8nButton,
@@ -56,6 +55,7 @@ import { useKeybindings } from '@/composables/useKeybindings';
 import { useCalloutHelpers } from '@/composables/useCalloutHelpers';
 import ProjectNavigation from '@/features/projects/components/ProjectNavigation.vue';
 import MainSidebarSourceControl from './MainSidebarSourceControl.vue';
+import MainSidebarUserArea from '@/components/MainSidebarUserArea.vue';
 
 const becomeTemplateCreatorStore = useBecomeTemplateCreatorStore();
 const cloudPlanStore = useCloudPlanStore();
@@ -83,23 +83,11 @@ useKeybindings({
 });
 
 // Template refs
-const user = ref<Element | null>(null);
+const user = useTemplateRef('user');
 
 // Component data
 const basePath = ref('');
 const fullyExpanded = ref(false);
-const userMenuItems = ref<IMenuItem[]>([
-	{
-		id: 'settings',
-		icon: 'settings',
-		label: i18n.baseText('settings'),
-	},
-	{
-		id: 'logout',
-		icon: 'door-open',
-		label: i18n.baseText('auth.signout'),
-	},
-]);
 
 const showWhatsNewNotification = computed(
 	() =>
@@ -326,9 +314,9 @@ const userIsTrialing = computed(() => cloudPlanStore.userIsTrialing);
 onMounted(async () => {
 	window.addEventListener('resize', onResize);
 	basePath.value = rootStore.baseUrl;
-	if (user.value) {
+	if (user.value?.$el) {
 		void externalHooks.run('mainSidebar.mounted', {
-			userRef: user.value,
+			userRef: user.value.$el,
 		});
 	}
 
@@ -347,23 +335,6 @@ const trackHelpItemClick = (itemType: string) => {
 		type: itemType,
 		workflow_id: workflowsStore.workflowId,
 	});
-};
-
-const onUserActionToggle = (action: string) => {
-	switch (action) {
-		case 'logout':
-			onLogout();
-			break;
-		case 'settings':
-			void router.push({ name: VIEWS.SETTINGS });
-			break;
-		default:
-			break;
-	}
-};
-
-const onLogout = () => {
-	void router.push({ name: VIEWS.SIGNOUT });
 };
 
 const toggleCollapse = () => {
@@ -624,54 +595,12 @@ onClickOutside(createBtn as Ref<VueInstance>, () => {
 		</N8nScrollArea>
 
 		<MainSidebarSourceControl :is-collapsed="isCollapsed" />
-		<div v-if="showUserArea">
-			<div ref="user" :class="$style.userArea">
-				<N8nPopoverReka side="right" align="end" :side-offset="16">
-					<template #content>
-						<div :class="$style.popover">
-							<N8nMenuItem
-								v-for="action in userMenuItems"
-								:key="action.id"
-								:item="action"
-								:data-test-id="`user-menu-item-${action.id}`"
-								@click="() => onUserActionToggle(action.id)"
-							/>
-						</div>
-					</template>
-					<template #trigger>
-						<div :class="$style.userAreaInner">
-							<div class="ml-3xs" data-test-id="main-sidebar-user-menu">
-								<!-- This dropdown is only enabled when sidebar is collapsed -->
-								<div :class="{ [$style.avatar]: true, ['clickable']: isCollapsed }">
-									<N8nAvatar
-										:first-name="usersStore.currentUser?.firstName"
-										:last-name="usersStore.currentUser?.lastName"
-										size="small"
-									/>
-								</div>
-							</div>
-							<div
-								:class="{
-									['ml-2xs']: true,
-									[$style.userName]: true,
-									[$style.expanded]: fullyExpanded,
-								}"
-							>
-								<N8nText size="small" color="text-dark">{{
-									usersStore.currentUser?.fullName
-								}}</N8nText>
-							</div>
-							<div
-								data-test-id="user-menu"
-								:class="{ [$style.userActions]: true, [$style.expanded]: fullyExpanded }"
-							>
-								<N8nIconButton icon="ellipsis" text square type="tertiary" />
-							</div>
-						</div>
-					</template>
-				</N8nPopoverReka>
-			</div>
-		</div>
+		<MainSidebarUserArea
+			v-if="showUserArea"
+			ref="user"
+			:fully-expanded="fullyExpanded"
+			:is-collapsed="isCollapsed"
+		/>
 
 		<TemplateTooltip />
 	</div>
@@ -755,44 +684,6 @@ onClickOutside(createBtn as Ref<VueInstance>, () => {
 .popoverTitle {
 	display: block;
 	margin-bottom: var(--spacing--3xs);
-}
-
-.userArea {
-	display: flex;
-	padding: var(--spacing--xs);
-	align-items: center;
-	border-top: var(--border-width) var(--border-style) var(--color--foreground);
-
-	.userName {
-		display: none;
-		overflow: hidden;
-		width: 100px;
-		white-space: nowrap;
-		text-overflow: ellipsis;
-
-		&.expanded {
-			display: initial;
-		}
-
-		span {
-			overflow: hidden;
-			text-overflow: ellipsis;
-		}
-	}
-
-	.userActions {
-		display: none;
-
-		&.expanded {
-			display: initial;
-		}
-	}
-}
-
-.userAreaInner {
-	display: flex;
-	align-items: center;
-	width: 100%;
 }
 
 @media screen and (max-height: 470px) {
